services/fax/src/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.*
google-cloud-storage==2.*
google-api-python-client==2.*
google-auth-httplib2==0.*
google-auth-oauthlib==1.*
